pyproject.toml
[tool.poetry]
name = "equirectangular_to_multi_perspective_generator"
version = "0.1.0"
description = ""
authors = ["meg-ssk"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.9"
scipy = "^1.8.1"
opencv-python = "^4.5.5"
PyYAML = "^6.0"
numpy = "^1.22.4"
py360convert = "^0.1.0"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
